comment verification

+++ src/main/java/life/majiang/community2/controller/QuestionController.java
@@ -22,7 +22,7 @@ public class QuestionController {
 
     @GetMapping("/question/{id}")
     public String question(Model model,
-                           @PathVariable(name = "id") Integer id) {
+                           @PathVariable(name = "id") Long id) {
         QuestionDTO questionDTO = questionService.getById(id);
 
         //累加阅读数
